Fix incorrect behaviour of the ToolbarLayoutListener.
ToolbarLayoutListener holds a WeakReference to the CalligraphyFactory.
In some cases (usually when we're trying to show a Toolbar for the first time)
referenced object can be null when onGlobalLayout() is called. As a result
the subtitle of the Toolbar will be " " (space character), the title
will be displayed at the top of the Toolbar, and custom fonts won't be applied.